fix: fix project structure
