+++ driverService/src/main/java/com/modsen/driverservice/exception/DuplicateFieldException.java
@@ -1,0 +1,11 @@
+package com.modsen.driverservice.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class DuplicateFieldException extends RuntimeException {
+    private String message;
+    private Long errorCode;
+}
