+++ driverService/src/main/java/com/modsen/driverservice/exception/ErrorMessage.java
@@ -1,0 +1,13 @@
+package com.modsen.driverservice.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ErrorMessage {
+    private Long errorCode;
+    private String errorMessage;
+}
