+++ driverService/src/main/java/com/modsen/driverservice/mapper/DriverMapper.java
@@ -1,0 +1,17 @@
+package com.modsen.driverservice.mapper;
+
+import com.modsen.driverservice.dto.DriverRequestDTO;
+import com.modsen.driverservice.dto.DriverResponseDTO;
+import com.modsen.driverservice.model.Driver;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface DriverMapper {
+    Driver toDriver(DriverRequestDTO driverRequestDTO);
+
+    Driver toDriver(DriverResponseDTO driverResponseDTO);
+
+    DriverRequestDTO toDriverRequestDTO(Driver driver);
+
+    DriverResponseDTO toDriverResponseDTO(Driver driver);
+}
