--- src/main/java/org/example/taxiprovider/TaxiProviderApplication.java
@@ -1,13 +1,0 @@
-package org.example.taxiprovider;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TaxiProviderApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TaxiProviderApplication.class, args);
-    }
-
-}
