--- src/services/driverService/src/main/java/org/example/driverservice/exception/ErrorMessage.java
@@ -1,13 +1,0 @@
-package org.example.driverservice.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ErrorMessage {
-    private Long errorCode;
-    private String errorMessage;
-}
