--- src/services/driverService/src/main/java/org/example/driverservice/exception/NotFoundException.java
@@ -1,11 +1,0 @@
-package org.example.driverservice.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class NotFoundException extends RuntimeException {
-    private String message;
-    private Long errorCode;
-}
