--- src/services/driverService/src/main/java/org/example/driverservice/mapper/DriverListMapper.java
@@ -1,15 +1,0 @@
-package org.example.driverservice.mapper;
-
-import org.example.driverservice.dto.DriverRequestDTO;
-import org.example.driverservice.dto.DriverResponseDTO;
-import org.example.driverservice.model.Driver;
-import org.mapstruct.Mapper;
-
-import java.util.List;
-
-@Mapper(componentModel = "spring", uses = DriverMapper.class)
-public interface DriverListMapper {
-    List<Driver> toDriverList(List<DriverRequestDTO> driverRequestDTOList);
-
-    List<DriverResponseDTO> toDriverResponseDTOList(List<Driver> driverList);
-}
