--- src/services/driverService/src/main/java/org/example/driverservice/service/DriverService.java
@@ -1,18 +1,0 @@
-package org.example.driverservice.service;
-
-import org.example.driverservice.dto.DriverRequestDTO;
-import org.example.driverservice.dto.DriverResponseDTO;
-
-import java.util.List;
-
-public interface DriverService {
-    List<DriverResponseDTO> getAllDrivers();
-
-    DriverResponseDTO getDriverById(Long id);
-
-    DriverResponseDTO createDriver(DriverRequestDTO driverRequestDTO);
-
-    DriverResponseDTO updateDriver(DriverRequestDTO driverRequestDTO);
-
-    void deleteDriver(Long id);
-}
