--- src/test/java/org/example/taxiprovider/TaxiProviderApplicationTests.java
@@ -1,13 +1,0 @@
-package org.example.taxiprovider;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TaxiProviderApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
